chore!: rework and simplify using TDD
I was not satisfied with unintuitive method names and the state of
tests. As pragmatic TDD focuses primarily on testing contracts with
the user (i.e. the module public API), it gave me an opportunity
to rethink it again, make it more user-friendly, with high
confidence in the new tests.

BREAKING CHANGE: The core functionality is still the same, only the
names have changed - now they are more meaningful. Functions for
retrieving data from transaction reports and account statements have
been converted to static methods on `fio_banka.Account`. For the
new API overview, see the module docstring.